Abstracting the Proposed block graph component
